OS sleep support
Created new interface for tasks to interact with scheduler
Made context switches less frequent (100Hz).
Tasks can now request to sleep for X milliseconds.
Stores the current scheduler time and wait time in the
task descriptor. Scheduler keeps track of the current time
in a static var and checks this against the wait start & duration
in the task descriptor to see if it should run that task.